Finish refactor, pre hook together
